Add polish.

+++ app/src/main/java/me/susieson/popularmovies/model/Movie.java
@@ -5,10 +5,10 @@ public class Movie {
     private String originalTitle;
     private String posterPath;
     private String overview;
-    private int voteAverage;
+    private double voteAverage;
     private String releaseDate;
 
-    public Movie(String originalTitle, String posterPath, String overview, int voteAverage, String releaseDate) {
+    public Movie(String originalTitle, String posterPath, String overview, double voteAverage, String releaseDate) {
         this.originalTitle = originalTitle;
         this.posterPath = posterPath;
         this.overview = overview;
@@ -28,7 +28,7 @@ public class Movie {
         return overview;
     }
 
-    public int getVoteAverage() {
+    public double getVoteAverage() {
         return voteAverage;
     }
